# swagger_ingilizce_app.yaml
openapi: 3.0.0
info:
  title: İngilizce Öğrenme Uygulaması API (Sosyal & İstatistikli)
  description: |
    Mobil uygulama için geliştirilmiş, öğrenme modülleri, detaylı istatistikler ve sosyal özellikler içeren REST API.
  version: 1.0.0

# --- DÜZELTME: Sunucu adresini kök yaptık (sonda /api yok) ---
servers:
  - url: http://127.0.0.1:8000
    description: Docker/Yerel Geliştirme Ortamı

# --- JWT Ayarları ---
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Kimlik Doğrulama]
      summary: Kullanıcı Girişi
      security: []  # Giriş yaparken token gerekmez
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, example: user1 }
                password: { type: string, example: pass1 }
      responses:
        '200': { description: Giriş başarılı, istatistikler ve son giriş tarihi döndürüldü. }
  /api/auth/register:
    post:
      tags: [Kimlik Doğrulama]
      summary: Yeni Kullanıcı Kaydı
      security: []  # Kayıt olurken token gerekmez
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, example: yeni_kullanici }
                password: { type: string, example: sifre123 }
      responses:
        '201': { description: Kayıt başarılı. }

  # --- Öğrenme Modülleri ---
...